pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "depnova"
version = "0.1.0"
description = "Comprehensive dependency inventory and CVE scanning across backend, frontend, and OS packages."
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Shivhare"}
]

dependencies = [
    "click>=8.1",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "beautifulsoup4>=4.12",
    "lxml>=5.0",
    "httpx>=0.27",
    "structlog>=24.0",
    "diskcache>=5.6",
    "packageurl-python>=0.15",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
]

[project.scripts]
depnova = "depnova.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["depnova*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
